requirements.txt
CouchDB==0.9
Django==1.6.2
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.21
South==0.8.4
Werkzeug==0.9.4
argparse==1.2.1
cartodb==0.7
distribute==0.6.24
dj-database-url==0.3.0
dj-static==0.0.5
django-bootstrap-themes==3.1.2
django-bootstrap3==4.4.0
django-dajax==0.9.2
django-dajaxice==0.6
django-excel-response==1.0
django-geoposition==0.1.5
django-mailchimp-v1.3==1.3.1
django-pure-pagination==0.2.1
django-suit==0.2.6
django-toolbelt==0.0.1
django-wpadmin==1.6.1
gunicorn==18.0
httplib2==0.8
itsdangerous==0.24
oauth2==1.5.211
prettytable==0.7.2
psycopg2==2.5.2
pycouchdb==1.7
pystache==0.5.3
python-dateutil==2.2
requests==2.2.1
simplejson==3.3.3
six==1.6.1
static==1.0.2
wsgiref==0.1.2
xlwt==0.7.5
